4.21 blogin voi poistaa vain sen lisännyt käyttäjä

+++ controllers/blogs.js
@@ -14,21 +14,9 @@ blogsRouter.get('/', async (request, response, next) => {
   }
 })
 
-/*
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization.toLowerCase().startsWith('bearer ')){
-    return authorization.substring(7)
-  }
-  return null
-}
-*/
-
 blogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-  //const token = getTokenFrom(request)
   try {
-    //const decodedToken = jwt.verify(token, process.env.SECRET)
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     if (!request.token || !decodedToken.id) {
       return response.status(401).json({error: 'token missing or invalid'})
@@ -55,6 +43,15 @@ blogsRouter.post('/', async (request, response, next) => {
 blogsRouter.delete('/:id', async (request, response, next) => {
   
   try {
+    //decodedToken: {username: <username>, id: <userid>}
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({error: 'token missing or invalid'})
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      response.status(401).json({error: 'unauthorized user'})
+    }
     const deletable = await Blog.findByIdAndRemove(request.params.id)
 
     if (deletable) {
